Wire up index/get droplets to database
The Droplets service now actually queries the database for either all
droplets or a specific droplet (by id). Two routes exist that correspond
with each of these queries. At the moment, no pre or post processing is
happening to the data -- it's just retrieved from the database and then
sent along.